requirements.txt
torch==2.2.2+cpu
torchvision==0.17.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
pillow
fastapi
uvicorn
matplotlib
scikit-learn
opencv-python